refactor(shorts): rename files to lowercase kebab-case
