+++ apps/shorts/src/components/technique-sheet.tsx
