+++ apps/shorts/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import App from "./App";
+import App from "./app";
 
 const root = document.getElementById("root");
 if (!root) throw new Error("Root element not found");
